requirements-full.txt
boto3>=1.34
click>=8.1
flake8>=7.0
GitPython>=3.1
openai>=1.30
pytest>=8.0
pydantic>=2.0
python-dotenv>=1.0
requests>=2.0
# unity-cli
pyyaml>=6.0
